Add Koji remote repo content consolidation to local hosted repo
This implements a backgrounded / threaded process whenever a new Koji remote repository is created,
which will proactively download all available files (based on the Koji build / archive records), and
copy that content into a single hosted repository, creating the target repo and adding it to the
group first if necessary. The target hosted repository will be added to the top of the target group's
membership, and will have the same name as the target group. Once all content for the remote repository
is transferred into the hosted repository, associated metadata paths will be removed from both the hosted
repo and the target group, so these will be regenerated. Finally, the newly-created Koji remote repository
will be removed from the group so as to keep the group as slim as possible.

Along with this new feature, the patch also adds a new repair endpoint for koji which will attempt to
go through all of the Koji remote repositories in the membership of some target group and run the same
consolidation logic on all of them, in the order in which they appear in the membership.

Consolidation will use checksums available in the Koji build records, along with those calculated during
various transfers (both on download and on copy to the consolidated repo) to ensure nothing is corrupted.

If content is corrupted, the download will be retried up to 3 times per path.

If the consolidation fails because of corruption or some other reason, the remote repo will be retained
but marked as having failed consolidation for later follow-up with some new type of repair feature.

This patch also contains logic to clear metadata caches when the corresponding metadata file is
deleted via the ContentManager component. This should help us de-cache incorrect metadata when it happens.