Add custom deserializer for VersionMetadata
Adding this aims to handle the parsing stuff more flexible I hope,
seems the rules of package metadata template is loose/open, I see
most of the repository of version metadata is a object containing
"type" and "url",  but we ran into the pakcage "tmp" which the repository
is a array, and that appears only in one of their versions(e.g.: 0.0.4).
Details see: http://registry.npmjs.org/tmp

I don't think we need to update our model to adapt this type of data,
but we have to consider it because it will break our code and cause
no versions for the package.